UpdateDatabaseApp
This is the most recent update of the MySQL and Axios using API as of 4/27/20.  It requires you also download the util.js, which contains the Create_IMDB_URL() and CreateUrl() functions in order for the program to work properly.If the Check_database_title() function does not work suddenly, you forgot to download the util.js and place it in the same place where your app.js (main program) is.  If you wish to unit test also download the util.test.js file and place util.test.js file into a  folder called 'test'.